removed unneeded components, only using PhysicalObject component
