--- packages/newtonxr/src/objects/PhysicalCube.tsx
@@ -1,11 +1,0 @@
-"use client";
-
-import React from "react";
-
-export function PhysicalCube() {
-  return (
-    <>
-      <></>
-    </>
-  );
-}
