+++ packages/newtonxr/src/objects/index.ts
@@ -1,3 +1,1 @@
-export * from "./PhysicalBall.js";
-export * from "./PhysicalCube.js";
 export * from "./PhysicalObject.js";
